WIP: Model CVE Policy as embedded JSON strings
I've been exploring embedding the policy JSON as a string into the HCL
as an alternative to modelling the structure in HCL.

It simplifies the client code but complicates the resource_cve_policy
and tests quite a bit.
The Twistlock API returns additional unused fields, and sets a modified
field in the responses you get from it, that means the JSON returned by
refreshing from Twistlock won't match the JSON in the Terraform state.
The possible options are:
1. put that responsibility onto the Terraform user and make them embed
   JSON that sets all the unused fields to default values, and update
   modified times themselves. I'm not keen on this because it makes
   mistakes easy and Twistlock silently drops invalid input.
2. process the returned JSON to strip out the unused fields.
    * this is most easily done by creating Go structs to model the JSON structure
    * even in this case, the round-tripped JSON string may not match
      what's in the Terraform state due to key ordering in maps.