ci: Don't print CodeCov token in build output
Remove the `--verbose` argument from `curl` so that the token doesn't
get outputted with the full URL. The command will still be printed, but
the variables aren't interpolated at that point. I have:

- regenerated the current token at: https://codecov.io/gh/circleci/terraform-provider-twistlock/settings
- updated the token at: https://circleci.com/gh/circleci/terraform-provider-twistlock/edit#env-vars

This is in preperation of the repo being made public so that
`circleci/terraform-docker-hub-image` can pull from it at build time
without any complicated deploy key setup. After merging this we should
delete the SSH user key and deploy keys.

I ran `trufflehog` v2.0.99 on the repo to check the history for other
secrets that shouldn't be made public. Ignoring the `vendor` directory
because it contains third-party code that is also public:

    trufflehog -x <(echo vendor) . | less -R

It only identified:

- test fixtures in `twistlock/testdata/test-gpg-keys/`
- high entropy constant string in `password/password.go`